db::commitlog_entry_serializer: define a commitlog_entry_writer_base
commitlog_entry_writer_base is a template from a class that has a static
function that performs the actual serialization of the corresponding object.

This class implements all generic operation and virtual methods the commitlog entry writer has
to implement.

The some_commitlog_writer class that may be given to commitlog_entry_writer_base as a template
parameter has to implement 2 methods:
   1) template <typename Output>
      static void do_serialize(Output&, commitlog_entry_writer_base<some_commitlog_writer>*)

      This method has to serialize the data into output stream Output.

   2) static size_t estimate_size(commitlog_entry_writer_base<serialize_one_commitlog_entry>*)

      This method returns an estimate for the entry size without the actual serialization.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>