build: add --target flag to override -march value
This is probably the simplest way to make the build work on other
architectures. --target can be set to an empty string to allow
the compiler's default to be used.

If --target is not set then the default is going to be 'nehalem' on
x86 machines and the compiler's default on all other platforms.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Signed-off-by: Michael Munday <mike.munday@ibm.com>