utils: introduce loading_shared_values
This class implements an key-value container that is populated
using the provided asynchronous callback.

The value is loaded when there are active references to the value for the given key.

Container ensures that only one entry is loaded per key at any given time.

The returned value is a lw_shared_ptr to the actual value.

The value for a specific key is immediately evicted when there are no
more references to it.

The container is based on the boost::intrusive::unordered_set and is rehashed if needed
every time a new value is added (asynchronously loaded) or removed from it (the last reference to
the value is destroyed).

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>