service::storage_service: join all (hopefully) mutate related flags into an enum_set
There are quite a few flags currently passed as "bool" parameters to WRITE path related
functions. So, instead of passing them as separate "bool"s pass them in a single enum_set.

This is going to add the visibility too (e.g. one will be able to grep a particular flag to find
where it's used).

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>