cql_transport::cql_server: update the _unpaged_queries of the proper cql_server instance
process_query() may run on the shard different from the one the cql_server
instance referenced by the connection::_server belongs to (it points to
the instance that belongs to the same shard the "connection" instance belongs to).

Let's make sure we update the statistics counter above for the cql_server
object on the shard where the processing is actually taking place.

To achieve this we'll keep the reference to the proper cql_server object
in the "request copy" of the client_state.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>