tracing: store the prepared statements parameters values
Store the prepared statement positional parameters values in the
corresponding system_traces.sessions entry in the 'parameters' column
(which has a map<text,text> type).

Parameters are stored as a pair of "paramX" : "value", where X is
the index of the parameter starting from 0 and the "value" is the first
64 bytes of the parameter value represented as a bigInt hex value.

If the actual value is greater than 64 bytes then the "value" will
contain only 64 most significant bytes of it and will have the "..." at
the end.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>