tracing: tracing: add add_metric()/remove_metric() methods
These methods control the set of metrics names which values are recorded
when a tracing event is recorded.

Build the set of metrics instances handles in add_metric(<metric family name>).
The metric handle will contain the full name and the reference to the metric
function so that the call at the event creation is as fast and lightweight as
possible.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>