sstables: get rid of #include <seastar/core/shared_ptr_incomplete.hh>
Including this header has unpleasant consequences when there are more than
one instantiations of the template function inside it - in this case
compiler requires all types that require this template to be complete
in all complication units where any of such types is required. As a result
this requires an enormous amount of pointless #include's.

Better have a corresponding specialization in the corresponding .cc file.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>