compact_query_state: add statistics to account the size of tombstones and data covered by tombestones
Account the size of read tombstones and data that is covered by (any)
tombstone.

Store these counters in a shared object because compact_query_state
is used with a move-semantics a lot and in such cases we need to get statistics
when the corresponding compact_query_state object is not available any more.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>