commitlog: introduce a bad_segment_error
Introduce a common base class for all errors that indicate that the current
segment is broken.

This allows a laconic "catch" clause for all such errors.

The new class has a what() method that returns either a string
returned by an std::exception::what() method of the derived class or
a default "Bad segment error" string if a derived class doesn't have an
std::exception base class.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>