gossip: inform remote nodes that the bootstrapping node goes DOWN
Inform other nodes in the cluster that the bootstrapping node is about
to shut itself down if bootstrapping failed for some reason.

There may still be scheduled write handlers that were created prior to
that and they have corresponding CL values that are not going to be
satisfied (because the bootstrapping node goes DOWN).

In order to prevent unnecessary "Write Timeout" errors let's go over
pending write requests, remove the bootstrapping node's IP from targets
and adjust the CL calculation values.

Fixes #4488

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>