gossip: inform remote nodes that the bootstrapping node goes DOWN
Inform other nodes in the cluster that the bootstrapping nodes is about
to shut itself down if bootstrapping failed for some reason.

There may still be scheduling write handlers that
were created prior to that and they have corresponding CL values that are
not going to be satisfied (because bootstrapping node goes DOWN).

In order to prevent unnecessary "Write Timeout" errors let's go over
pending write requests, remove the corresponding IP from targets and
adjust the CL calculation values.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>