storage_proxy + gossiper: cleanup pending write handlers when bootstrapping node goes DOWN
If bootstrapping failed and the node gets and indication that the bootstrapping
node is going DOWN there may still be scheduling write handlers that
were created prior to that and have corresponding CL values that are
not going to be satisfied.

In order to prevent unnecessary "Write Timeout" errors let's go over
pending write requests, remove the corresponding IP from targets and
adjust the CL calculation values.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>