endpoint_state: replicate the 'is_alive' state when replicating the endpoint state across shards
is_alive is a part of the state hence there is no reason not to replicate
it with the rest of the state to begin with.

Without this patch it takes 1s (a gossiper round period) for an UP/DOWN
information to propagate from shard0 to other shards.

During this time we are going to be wasting precious CPU cycles trying to
access the already DOWN node.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>