configure.py: ignore 'thrift -version' exit code
(At least) on Ubuntu 19 'thrift -version' prints the expected
string but its exit status is non-zero:

$ thrift -version
Thrift version 0.9.1
$ echo $?
1

We don't really care about the exit status but rather about the printed
version string. If there is going to be some problem with the command,
e.g. it's missing, the printed string is not going to be as expected
anyway.

Therefore it should be safe to ignore the exit status.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>