hinted handoff: protect deletion of per-destination elements with a semaphore
When elements of manager::_ep_managers map or endpoints' directories are
deleted we want to serialize such events with other asynchronous parts of the code
that may be using these elements or rely on existence of these directories.

There are a few races that may occur if these events are not serialized:
 - #4685: space watchdog takes a per end_point_hints_manager shared_lock
   and if the corresponding end_point_hints_manager object is being destroyed
   while space watchdog is waiting for this lock there's going to be a
   use-after-free event.
 - Space watchdog scans for per-destination directories and if it's deleted
   while it does that it may lead to an error.
 - If drain_for() is running together with itself: one instance for the local
   node and one for some other node, erasing of elements from the _ep_managers
   map may again lead to a use-after-free event.

In this patch we introduce a semaphore that is going to serialize the context
where such deletions can take place and the context that can be accessing
these elements during the deletion.

Fixes #4685

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>